Dont override existing files

+++ init.py
@@ -40,11 +40,12 @@ class AdventManager:
     def _setup_code_templates(self, year: int, day: int) -> None:
         code = pathlib.Path(str(year), str(day), "code.py")
         code.parent.mkdir(parents=True, exist_ok=True)
-        with code.open("w", encoding="utf-8") as c:
-            c.write(
-                f'# Advent of Code | Year {year} | Day {str(day)} \n # Author {AUTHOR}\n\nwith open((__file__.rstrip("code.py")+"input.txt"), \'r\') as input_file:\n    input = input_file.read()\n\n\n\nprint("Part One : "+ str(None))\n\n\n\nprint("Part Two : "+ str(None))\n\n\nclass Solver:\n    def __init__(self, input):\n        self.input = input\n\n\nif __name__ == "__main__":\n    s = Solver()\n    s.run(input=input)'
-            )
-        pathlib.Path(str(year), str(day), "test.py").touch()
+        if not code.is_file():
+            with code.open("w", encoding="utf-8") as c:
+                c.write(
+                    f'# Advent of Code | Year {year} | Day {str(day)} \n # Author {AUTHOR}\n\nwith open((__file__.rstrip("code.py")+"input.txt"), \'r\') as input_file:\n    input = input_file.read()\n\n\n\nprint("Part One : "+ str(None))\n\n\n\nprint("Part Two : "+ str(None))\n\n\nclass Solver:\n    def __init__(self, input):\n        self.input = input\n\n\nif __name__ == "__main__":\n    s = Solver()\n    s.run(input=input)'
+                )
+            pathlib.Path(str(year), str(day), "test.py").touch()
         return
 
     def _collect_challenges(self, year: int, day: int) -> None:
